worker/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
playwright==1.47.0
Pillow==10.4.0
gspread==6.1.4
google-auth==2.35.0
google-api-python-client==2.149.0
python-dotenv==1.0.1
